MSVC compliance and Wwise integration start

+++ gui/Image.cpp
@@ -24,22 +24,22 @@ Point2i GUI::Image::gDimensions() {
 void GUI::Image::draw() {
   if(tex && !clip.empty()) {
     Vector<float> vertex(8), texCoord(8);
-    vertex[0] = clip.min().x();
-    vertex[1] = clip.min().y();
-    vertex[2] = clip.max().x();
-    vertex[3] = clip.min().y();
-    vertex[4] = clip.max().x();
-    vertex[5] = clip.max().y();
-    vertex[6] = clip.min().x();
-    vertex[7] = clip.max().y();
-    texCoord[0] = (clip.min().x()>pos.x()) ? (float)(clip.min().x()-pos.x())/dimensions.x() : 0.0;
-    texCoord[1] = (clip.min().y()>pos.y()) ? (float)(clip.min().y()-pos.y())/dimensions.y() : 0.0;
-    texCoord[2] = (clip.max().x()<pos.x()+dimensions.x()) ? (float)(clip.max().x()-pos.x())/dimensions.x() : 1.0;
-    texCoord[3] = (clip.min().y()>pos.y()) ? (float)(clip.min().y()-pos.y())/dimensions.y() : 0.0;
-    texCoord[4] = (clip.max().x()<pos.x()+dimensions.x()) ? (float)(clip.max().x()-pos.x())/dimensions.x() : 1.0;
-    texCoord[5] = (clip.max().y()<pos.y()+dimensions.y()) ? (float)(clip.max().y()-pos.y())/dimensions.y() : 1.0;
-    texCoord[6] = (clip.min().x()>pos.x()) ? (float)(clip.min().x()-pos.x())/dimensions.x() : 0.0;
-    texCoord[7] = (clip.max().y()<pos.y()+dimensions.y()) ? (float)(clip.max().y()-pos.y())/dimensions.y() : 1.0;
+    vertex[0] = (float)clip.min().x();
+    vertex[1] = (float)clip.min().y();
+    vertex[2] = (float)clip.max().x();
+    vertex[3] = (float)clip.min().y();
+    vertex[4] = (float)clip.max().x();
+    vertex[5] = (float)clip.max().y();
+    vertex[6] = (float)clip.min().x();
+    vertex[7] = (float)clip.max().y();
+    texCoord[0] = (clip.min().x()>pos.x()) ? (float)(clip.min().x()-pos.x())/dimensions.x() : 0.f;
+    texCoord[1] = (clip.min().y()>pos.y()) ? (float)(clip.min().y()-pos.y())/dimensions.y() : 0.f;
+    texCoord[2] = (clip.max().x()<pos.x()+dimensions.x()) ? (float)(clip.max().x()-pos.x())/dimensions.x() : 1.f;
+    texCoord[3] = (clip.min().y()>pos.y()) ? (float)(clip.min().y()-pos.y())/dimensions.y() : 0.f;
+    texCoord[4] = (clip.max().x()<pos.x()+dimensions.x()) ? (float)(clip.max().x()-pos.x())/dimensions.x() : 1.f;
+    texCoord[5] = (clip.max().y()<pos.y()+dimensions.y()) ? (float)(clip.max().y()-pos.y())/dimensions.y() : 1.f;
+    texCoord[6] = (clip.min().x()>pos.x()) ? (float)(clip.min().x()-pos.x())/dimensions.x() : 0.f;
+    texCoord[7] = (clip.max().y()<pos.y()+dimensions.y()) ? (float)(clip.max().y()-pos.y())/dimensions.y() : 1.f;
     GL::Utils::draw2dTexQuad(vertex,texCoord,*tex);
   }
 }
